+++ auth_service/src/test/java/com/nicovegasr/auth_service/unit_test/models/user/UserTest.java
@@ -1,0 +1,6 @@
+package com.nicovegasr.auth_service.unit_test.models.user;
+
+public class UserTest {
+    /** Use cases:
+     */
+}
